refactor: decouple demos from job_scheduling
