+++ CMakeLists.txt
@@ -52,6 +52,7 @@ set(SRC_DIRS
     demos/bit_manipulation
     demos/error_correction_algorithms
     demos/hashing
+    demos/job_scheduling
     features/telemetry
     features/serialization
     features/dp_visualizer
