+++ demos/job_scheduling/priority_scheduling_demo.c
@@ -1,0 +1,32 @@
+#include "job_scheduling.h"
+#include <stdio.h>
+#include <time.h>
+
+void priority_scheduling_demo(void)
+{
+    Process procs[10];
+    int n;
+
+    if (!js_read_processes(procs, &n, 1))
+    {
+        printf("\nExiting priority scheduling demo....\n");
+        return;
+    }
+
+    clock_t start_t, end_t;
+    double total_t;
+
+    start_t = clock();
+
+    GanttSegment segments[JS_MAX_SEGMENTS];
+    int segment_count = 0;
+    
+    priority_scheduling(procs, n, segments, &segment_count);
+
+    end_t = clock();
+    total_t = (double)(end_t - start_t) / CLOCKS_PER_SEC;
+
+    js_print_result(procs, n);
+    js_print_gantt(segments, segment_count);
+    printf("\ntotal CPU time taken for priority scheduling:- %f seconds\n", total_t);
+}
