+++ src/job_scheduling/fcfs.c
@@ -9,22 +9,9 @@
 // starts after the process list is read and stops before the results are
 // printed). it is for demonstration only and must not be treated as a measure
 // of the algorithm's efficiency.
-void fcfs_demo(void)
+
+void fcfs_schedule(Process *procs, int n, GanttSegment *segments, int *segment_count)
 {
-    Process procs[10];
-    int n;
-
-    if (!js_read_processes(procs, &n, 0))
-    {
-        printf("\nExiting FCFS demo....\n");
-        return;
-    }
-
-    clock_t start_t, end_t;
-    double total_t;
-
-    start_t = clock();
-
     // Stable insertion sort to keep input order on ties
     for (int i = 1; i < n; i++)
     {
@@ -38,8 +25,6 @@ void fcfs_demo(void)
         procs[j + 1] = key;
     }
 
-    GanttSegment segments[JS_MAX_SEGMENTS];
-    int segment_count = 0;
     int current_time = 0;
 
     for (int i = 0; i < n; i++)
@@ -47,13 +32,13 @@ void fcfs_demo(void)
         // idle until this process arrives
         while (current_time < procs[i].arrival)
         {
-            js_add_segment(segments, &segment_count, -1, current_time);
+            js_add_segment(segments, segment_count, -1, current_time);
             current_time++;
         }
 
         for (int t = 0; t < procs[i].burst; t++)
         {
-            js_add_segment(segments, &segment_count, procs[i].id, current_time);
+            js_add_segment(segments, segment_count, procs[i].id, current_time);
             current_time++;
         }
 
@@ -61,11 +46,4 @@ void fcfs_demo(void)
         procs[i].turnaround = procs[i].completion - procs[i].arrival;
         procs[i].waiting = procs[i].turnaround - procs[i].burst;
     }
-
-    end_t = clock();
-    total_t = (double)(end_t - start_t) / CLOCKS_PER_SEC;
-
-    js_print_result(procs, n);
-    js_print_gantt(segments, segment_count);
-    printf("\ntotal CPU time taken for FCFS scheduling:- %f seconds\n", total_t);
 }
