+++ src/job_scheduling/preemptive_priority.c
@@ -9,24 +9,9 @@
 // preempt the running job.
 // note: the time measured by clock() covers the scheduling computation only and
 // is for demonstration only, not a measure of the algorithm's efficiency.
-void preemptive_priority_demo(void)
+
+void preemptive_priority_schedule(Process *procs, int n, GanttSegment *segments, int *segment_count)
 {
-    Process procs[10];
-    int n;
-
-    if (!js_read_processes(procs, &n, 1))
-    {
-        printf("\nExiting preemptive priority demo....\n");
-        return;
-    }
-
-    clock_t start_t, end_t;
-    double total_t;
-
-    start_t = clock();
-
-    GanttSegment segments[JS_MAX_SEGMENTS];
-    int segment_count = 0;
     int completed = 0;
     int current_time = 0;
 
@@ -52,12 +37,12 @@ void preemptive_priority_demo(void)
         if (chosen == -1)
         {
             // nothing has arrived yet, idle one tick
-            js_add_segment(segments, &segment_count, -1, current_time);
+            js_add_segment(segments, segment_count, -1, current_time);
             current_time++;
             continue;
         }
 
-        js_add_segment(segments, &segment_count, procs[chosen].id, current_time);
+        js_add_segment(segments, segment_count, procs[chosen].id, current_time);
         procs[chosen].remaining--;
         current_time++;
 
@@ -69,11 +54,4 @@ void preemptive_priority_demo(void)
             completed++;
         }
     }
-
-    end_t = clock();
-    total_t = (double)(end_t - start_t) / CLOCKS_PER_SEC;
-
-    js_print_result(procs, n);
-    js_print_gantt(segments, segment_count);
-    printf("\ntotal CPU time taken for preemptive priority scheduling:- %f seconds\n", total_t);
 }
