+++ src/job_scheduling/sjf.c
@@ -6,24 +6,9 @@
 // among those that have already arrived; once started it runs to completion.
 // note: the time measured by clock() covers the scheduling computation only and
 // is for demonstration only, not a measure of the algorithm's efficiency.
-void sjf_demo(void)
+
+void sjf_schedule(Process *procs, int n, GanttSegment *segments, int *segment_count)
 {
-    Process procs[10];
-    int n;
-
-    if (!js_read_processes(procs, &n, 0))
-    {
-        printf("\nExiting SJF demo....\n");
-        return;
-    }
-
-    clock_t start_t, end_t;
-    double total_t;
-
-    start_t = clock();
-
-    GanttSegment segments[JS_MAX_SEGMENTS];
-    int segment_count = 0;
     int done[10] = {0};
     int completed = 0;
     int current_time = 0;
@@ -49,14 +34,14 @@ void sjf_demo(void)
         if (chosen == -1)
         {
             // no job has arrived yet, idle one tick and try again
-            js_add_segment(segments, &segment_count, -1, current_time);
+            js_add_segment(segments, segment_count, -1, current_time);
             current_time++;
             continue;
         }
 
         for (int t = 0; t < procs[chosen].burst; t++)
         {
-            js_add_segment(segments, &segment_count, procs[chosen].id, current_time);
+            js_add_segment(segments, segment_count, procs[chosen].id, current_time);
             current_time++;
         }
 
@@ -66,11 +51,4 @@ void sjf_demo(void)
         done[chosen] = 1;
         completed++;
     }
-
-    end_t = clock();
-    total_t = (double)(end_t - start_t) / CLOCKS_PER_SEC;
-
-    js_print_result(procs, n);
-    js_print_gantt(segments, segment_count);
-    printf("\ntotal CPU time taken for SJF scheduling:- %f seconds\n", total_t);
 }
